Organize image_checker code
