+++ tests/integration/image_checker.py
@@ -1,0 +1,341 @@
+import os
+import shutil
+from math import ceil
+from typing import Dict, List
+
+import matplotlib.backends.backend_pdf
+import matplotlib.image as mpimg
+from mache import MachineInfo
+from matplotlib import pyplot as plt
+from PIL import Image, ImageChops, ImageDraw
+
+
+# Classes #####################################################################
+class Parameters(object):
+    def __init__(self, args: Dict[str, str]):
+        self.actual_images_dir = args["actual_images_dir"]
+        self.expected_images_dir = args["expected_images_dir"]
+        self.diff_dir = args["diff_dir"]
+        self.expected_images_list = args["expected_images_list"]
+
+
+class Results(object):
+    def __init__(
+        self,
+        diff_dir: str,
+        prefix: str,
+        image_count_total: int,
+        file_list_missing: List[str],
+        file_list_mismatched: List[str],
+    ):
+        if image_count_total == 0:
+            raise ValueError(f"No images found for task {prefix} in {diff_dir}")
+        self.diff_dir = diff_dir
+        self.prefix = prefix
+        self.image_count_total = image_count_total
+        self.image_count_missing = len(file_list_missing)
+        self.image_count_mismatched = len(file_list_mismatched)
+        self.image_count_correct = (
+            image_count_total - len(file_list_missing) - len(file_list_mismatched)
+        )
+        self.file_list_missing = sorted(file_list_missing)
+        self.file_list_mismatched = sorted(file_list_mismatched)
+
+
+# Specialized setup ###########################################################
+
+
+def set_up_and_run_image_checker(
+    cfg_specifier: str,
+    case_name: str,
+    expansions: Dict,
+    diff_dir_suffix: str,
+    task_list: List[str],
+    test_results_dict: Dict[str, Results],
+):
+    print(f"Image checking {cfg_specifier}")
+    actual_images_dir = f"{expansions['user_www']}zppy_weekly_{cfg_specifier}_www/{expansions['unique_id']}/{case_name}/"
+    d: Dict[str, str] = {
+        "actual_images_dir": actual_images_dir,
+        "expected_images_dir": f"{expansions['expected_dir']}expected_{cfg_specifier}",
+        "diff_dir": f"{actual_images_dir}image_check_failures_{cfg_specifier}{diff_dir_suffix}",
+        "expected_images_list": f"{expansions['expected_dir']}image_list_expected_{cfg_specifier}.txt",
+    }
+    parameters: Parameters = Parameters(d)
+    for task in task_list:
+        test_results = check_images(parameters, task)
+        test_results_dict[f"{cfg_specifier}_{task}"] = test_results
+
+
+# Everything below here could, in theory, be pulled out into an Image Checker package
+# (or more likely, a zppy-interfaces entry point)
+# Generalized image checking ##################################################
+
+
+def check_images(parameters: Parameters, prefix: str):
+    test_results = _check_mismatched_images(parameters, prefix)
+    diff_subdir = f"{parameters.diff_dir}/{prefix}"
+    if os.path.exists(diff_subdir):
+        # Write missing and mismatched images to files
+        missing_images_file = f"{diff_subdir}/missing_images.txt"
+        if os.path.exists(missing_images_file):
+            os.remove(missing_images_file)
+        for missing_image in test_results.file_list_missing:
+            with open(missing_images_file, "a") as f:
+                f.write(f"{missing_image}\n")
+        mismatched_images_file = f"{diff_subdir}/mismatched_images.txt"
+        if os.path.exists(mismatched_images_file):
+            os.remove(mismatched_images_file)
+        for mismatched_image in test_results.file_list_mismatched:
+            with open(mismatched_images_file, "a") as f:
+                f.write(f"{mismatched_image}\n")
+        # Create image diff grid
+        _make_image_diff_grid(diff_subdir)
+    else:
+        print(f"Diff subdir {diff_subdir} does not exist.")
+    return test_results
+
+
+def construct_markdown_summary_table(
+    test_results_dict: Dict[str, Results], output_file_path: str
+):
+    with open(output_file_path, "w") as f:
+        f.write("# Summary of test results\n\n")
+        f.write(
+            "Diff subdir is where to find the lists of missing/mismatched images, the image diff grid, and the individual diffs.\n"
+        )
+        f.write("Note image diff grids can not yet be constructed automatically.\n")
+        f.write(
+            "| Test name | Total images | Correct images | Missing images | Mismatched images | Diff subdir | \n"
+        )
+        f.write("| --- | --- | --- | --- | --- | --- | \n")
+        for test_name, test_results in test_results_dict.items():
+            f.write(
+                f"| {test_name} | {test_results.image_count_total} | {test_results.image_count_correct} | {test_results.image_count_missing} | {test_results.image_count_mismatched} | {test_results.diff_dir}/{test_results.prefix} | \n"
+            )
+    print(f"Copy the output of {output_file_path} to a Pull Request comment")
+
+
+# Helper functions ############################################################
+
+
+def _check_mismatched_images(
+    parameters: Parameters,
+    prefix: str,
+) -> Results:
+    missing_images: List[str] = []
+    mismatched_images: List[str] = []
+
+    counter = 0
+    print(f"Opening expected images file {parameters.expected_images_list}")
+    with open(parameters.expected_images_list) as f:
+        print(f"Reading expected images file {parameters.expected_images_list}")
+        for line in f:
+            image_name = line.strip("./").strip("\n")
+            proceed = False
+            if image_name.startswith(prefix):
+                proceed = True
+            if proceed:
+                counter += 1
+                if counter % 250 == 0:
+                    print("On line #", counter)
+                path_to_actual_png = os.path.join(
+                    parameters.actual_images_dir, image_name
+                )
+                path_to_expected_png = os.path.join(
+                    parameters.expected_images_dir, image_name
+                )
+
+                _compare_actual_and_expected(
+                    missing_images,
+                    mismatched_images,
+                    image_name,
+                    path_to_actual_png,
+                    path_to_expected_png,
+                    parameters.diff_dir,
+                )
+
+    if missing_images:
+        print("Missing images:")
+        for i in missing_images:
+            print(i)
+    if mismatched_images:
+        print("Mismatched images:")
+        for i in mismatched_images:
+            print(i)
+
+    # Count summary
+    print(f"Total: {counter}")
+    print(f"Number of missing images: {len(missing_images)}")
+    print(f"Number of mismatched images: {len(mismatched_images)}")
+    print(
+        f"Number of correct images: {counter - len(missing_images) - len(mismatched_images)}"
+    )
+    test_results = Results(
+        parameters.diff_dir, prefix, counter, missing_images, mismatched_images
+    )
+
+    # Make diff_dir readable
+    if os.path.exists(parameters.diff_dir):
+        os.system(f"chmod -R 755 {parameters.diff_dir}")
+    else:
+        # diff_dir won't exist if all the expected images are missing
+        # That is, if we're in this case, we expect the following:
+        assert len(missing_images) == counter
+
+    return test_results
+
+
+def _compare_actual_and_expected(
+    missing_images,
+    mismatched_images,
+    image_name,
+    path_to_actual_png,
+    path_to_expected_png,
+    diff_dir,
+):
+    # https://stackoverflow.com/questions/35176639/compare-images-python-pil
+    try:
+        actual_png = Image.open(path_to_actual_png).convert("RGB")
+    except FileNotFoundError:
+        missing_images.append(image_name)
+        return
+    expected_png = Image.open(path_to_expected_png).convert("RGB")
+    diff = ImageChops.difference(actual_png, expected_png)
+
+    if not os.path.isdir(diff_dir):
+        os.mkdir(diff_dir)
+
+    bbox = diff.getbbox()
+    if not bbox:
+        # If `diff.getbbox()` is None, then the images are in theory equal
+        assert diff.getbbox() is None
+    else:
+        # Sometimes, a few pixels will differ, but the two images appear identical.
+        # https://codereview.stackexchange.com/questions/55902/fastest-way-to-count-non-zero-pixels-using-python-and-pillow
+        nonzero_pixels = (
+            diff.crop(bbox)
+            .point(lambda x: 255 if x else 0)
+            .convert("L")
+            .point(bool)
+            .getdata()
+        )
+        num_nonzero_pixels = sum(nonzero_pixels)
+        width, height = expected_png.size
+        num_pixels = width * height
+        fraction = num_nonzero_pixels / num_pixels
+        # Fraction of mismatched pixels should be less than 0.02%
+        if fraction >= 0.0002:
+            verbose = False
+            if verbose:
+                print("\npath_to_actual_png={}".format(path_to_actual_png))
+                print("path_to_expected_png={}".format(path_to_expected_png))
+                print("diff has {} nonzero pixels.".format(num_nonzero_pixels))
+                print("total number of pixels={}".format(num_pixels))
+                print("num_nonzero_pixels/num_pixels fraction={}".format(fraction))
+
+            mismatched_images.append(image_name)
+
+            diff_dir_actual_png = os.path.join(
+                diff_dir, "{}_actual.png".format(image_name)
+            )
+            # image_name could contain a number of subdirectories.
+            os.makedirs(os.path.dirname(diff_dir_actual_png), exist_ok=True)
+            shutil.copy(
+                path_to_actual_png,
+                diff_dir_actual_png,
+            )
+            diff_dir_expected_png = os.path.join(
+                diff_dir, "{}_expected.png".format(image_name)
+            )
+            # image_name could contain a number of subdirectories.
+            os.makedirs(os.path.dirname(diff_dir_expected_png), exist_ok=True)
+            shutil.copy(
+                path_to_expected_png,
+                diff_dir_expected_png,
+            )
+            # Draw red box around diff-area on each of: diff, actual, expected
+            _draw_box(diff, diff, os.path.join(diff_dir, f"{image_name}_diff.png"))
+            _draw_box(
+                actual_png, diff, os.path.join(diff_dir, f"{image_name}_actual.png")
+            )
+            _draw_box(
+                expected_png, diff, os.path.join(diff_dir, f"{image_name}_expected.png")
+            )
+
+
+def _draw_box(image, diff, output_path: str):
+    # https://stackoverflow.com/questions/41405632/draw-a-rectangle-and-a-text-in-it-using-pil
+    draw = ImageDraw.Draw(image)
+    (left, upper, right, lower) = (
+        diff.getbbox()
+    )  # We specifically want the diff's bounding box
+    draw.rectangle(((left, upper), (right, lower)), outline="red")
+    image.save(output_path, "PNG")
+
+
+# TODO: fix issue where blank plots generate after so many pages in the PDF
+def _make_image_diff_grid(diff_subdir, pdf_name="image_diff_grid.pdf", rows_per_page=2):
+    machine_info = MachineInfo()
+    web_portal_base_path = machine_info.config.get("web_portal", "base_path")
+    web_portal_base_url = machine_info.config.get("web_portal", "base_url")
+    print(f"web_portal_base_path: {web_portal_base_path}")
+    print(f"web_portal_base_url: {web_portal_base_url}")
+    print(f"Making image diff grid for {diff_subdir}")
+
+    if not diff_subdir.startswith(web_portal_base_path):
+        print(
+            f"diff_subdir {diff_subdir} is not a subdir of web_portal_base_path: {web_portal_base_path}"
+        )
+        return
+    pdf_path = f"{diff_subdir}/{pdf_name}"
+    pdf = matplotlib.backends.backend_pdf.PdfPages(pdf_path)
+    print(f"Saving to:\n{pdf_path}")
+    web_subdir = diff_subdir.removeprefix(web_portal_base_path)
+    print(f"Web page will be at:\n{web_portal_base_url}/{web_subdir}/{pdf_name}")
+
+    prefixes = []
+    # print(f"Walking diff_subdir: {diff_subdir}")
+    for root, _, files in os.walk(diff_subdir):
+        # print(f"root: {root}")
+        for file_name in files:
+            # print(f"file_name: {file_name}")
+            if file_name.endswith("_diff.png"):
+                prefixes.append(f"{root}/{file_name.split('_diff.png')[0]}")
+    rows = len(prefixes)
+    if rows == 0:
+        # No diffs to collect into a PDF
+        return
+    cols = 3  # actual, expected, diff
+    print(f"Constructing a {rows}x{cols} grid of image diffs")
+
+    num_pages = ceil(rows / rows_per_page)
+    for page in range(num_pages):
+        fig, axes = plt.subplots(rows_per_page, cols)
+        print(f"Page {page}")
+        for i, ax_row in enumerate(axes):
+            count = page * 3 + i
+            if count > len(prefixes) - 1:
+                break
+            # We already know all the files are in `diff_subdir`; no need to repeat it.
+            short_title = prefixes[count].removeprefix(diff_subdir)
+            print(f"short_title {i}: {short_title}")
+            ax_row[1].set_title(short_title, fontsize=6)
+            img = mpimg.imread(f"{prefixes[count]}_actual.png")
+            ax_row[0].imshow(img)
+            ax_row[0].set_xticks([])
+            ax_row[0].set_yticks([])
+            img = mpimg.imread(f"{prefixes[count]}_expected.png")
+            ax_row[1].imshow(img)
+            ax_row[1].set_xticks([])
+            ax_row[1].set_yticks([])
+            img = mpimg.imread(f"{prefixes[count]}_diff.png")
+            ax_row[2].imshow(img)
+            ax_row[2].set_xticks([])
+            ax_row[2].set_yticks([])
+        fig.tight_layout()
+        pdf.savefig(1)
+        plt.close(fig)
+    pdf.close()
+    plt.close("all")
+    print(f"Reminder:\n{web_portal_base_url}/{web_subdir}/{pdf_name}")
