+++ tests/integration/utils.py
@@ -1,187 +1,11 @@
-import os
 import re
-import shutil
 import subprocess
-from typing import List
 
 from mache import MachineInfo
-from PIL import Image, ImageChops, ImageDraw
 
 UNIQUE_ID = "unique_id"
 
-# Image checking ##########################################################
-
-
-# Originally in https://github.com/E3SM-Project/zppy/pull/695
-# And https://github.com/E3SM-Project/zppy/pull/698
-class Results(object):
-    def __init__(
-        self,
-        diff_dir: str,
-        task: str,
-        image_count_total: int,
-        file_list_missing: List[str],
-        file_list_mismatched: List[str],
-    ):
-        if image_count_total == 0:
-            raise ValueError(f"No images found for task {task} in {diff_dir}")
-        self.diff_dir = diff_dir
-        self.task = task
-        self.image_count_total = image_count_total
-        self.image_count_missing = len(file_list_missing)
-        self.image_count_mismatched = len(file_list_mismatched)
-        self.image_count_correct = (
-            image_count_total - len(file_list_missing) - len(file_list_mismatched)
-        )
-        self.file_list_missing = sorted(file_list_missing)
-        self.file_list_mismatched = sorted(file_list_mismatched)
-
-
-# Copied from E3SM Diags
-def compare_images(
-    missing_images,
-    mismatched_images,
-    image_name,
-    path_to_actual_png,
-    path_to_expected_png,
-    diff_dir,
-):
-    # https://stackoverflow.com/questions/35176639/compare-images-python-pil
-    try:
-        actual_png = Image.open(path_to_actual_png).convert("RGB")
-    except FileNotFoundError:
-        missing_images.append(image_name)
-        return
-    expected_png = Image.open(path_to_expected_png).convert("RGB")
-    diff = ImageChops.difference(actual_png, expected_png)
-
-    if not os.path.isdir(diff_dir):
-        os.mkdir(diff_dir)
-
-    bbox = diff.getbbox()
-    if not bbox:
-        # If `diff.getbbox()` is None, then the images are in theory equal
-        assert diff.getbbox() is None
-    else:
-        # Sometimes, a few pixels will differ, but the two images appear identical.
-        # https://codereview.stackexchange.com/questions/55902/fastest-way-to-count-non-zero-pixels-using-python-and-pillow
-        nonzero_pixels = (
-            diff.crop(bbox)
-            .point(lambda x: 255 if x else 0)
-            .convert("L")
-            .point(bool)
-            .getdata()
-        )
-        num_nonzero_pixels = sum(nonzero_pixels)
-        width, height = expected_png.size
-        num_pixels = width * height
-        fraction = num_nonzero_pixels / num_pixels
-        # Fraction of mismatched pixels should be less than 0.02%
-        if fraction >= 0.0002:
-            verbose = False
-            if verbose:
-                print("\npath_to_actual_png={}".format(path_to_actual_png))
-                print("path_to_expected_png={}".format(path_to_expected_png))
-                print("diff has {} nonzero pixels.".format(num_nonzero_pixels))
-                print("total number of pixels={}".format(num_pixels))
-                print("num_nonzero_pixels/num_pixels fraction={}".format(fraction))
-
-            mismatched_images.append(image_name)
-
-            diff_dir_actual_png = os.path.join(
-                diff_dir, "{}_actual.png".format(image_name)
-            )
-            # image_name could contain a number of subdirectories.
-            os.makedirs(os.path.dirname(diff_dir_actual_png), exist_ok=True)
-            shutil.copy(
-                path_to_actual_png,
-                diff_dir_actual_png,
-            )
-            diff_dir_expected_png = os.path.join(
-                diff_dir, "{}_expected.png".format(image_name)
-            )
-            # image_name could contain a number of subdirectories.
-            os.makedirs(os.path.dirname(diff_dir_expected_png), exist_ok=True)
-            shutil.copy(
-                path_to_expected_png,
-                diff_dir_expected_png,
-            )
-            # https://stackoverflow.com/questions/41405632/draw-a-rectangle-and-a-text-in-it-using-pil
-            draw = ImageDraw.Draw(diff)
-            (left, upper, right, lower) = diff.getbbox()
-            draw.rectangle(((left, upper), (right, lower)), outline="red")
-            diff.save(
-                os.path.join(diff_dir, "{}_diff.png".format(image_name)),
-                "PNG",
-            )
-
-
-def check_mismatched_images(
-    actual_images_dir: str,
-    expected_images_file: str,
-    expected_images_dir: str,
-    diff_dir: str,
-    task: str,
-) -> Results:
-    missing_images: List[str] = []
-    mismatched_images: List[str] = []
-
-    counter = 0
-    print(f"Opening expected images file {expected_images_file}")
-    with open(expected_images_file) as f:
-        print(f"Reading expected images file {expected_images_file}")
-        for line in f:
-            image_name = line.strip("./").strip("\n")
-            proceed = False
-            if image_name.startswith(task):
-                proceed = True
-            if proceed:
-                counter += 1
-                if counter % 250 == 0:
-                    print("On line #", counter)
-                path_to_actual_png = os.path.join(actual_images_dir, image_name)
-                path_to_expected_png = os.path.join(expected_images_dir, image_name)
-
-                compare_images(
-                    missing_images,
-                    mismatched_images,
-                    image_name,
-                    path_to_actual_png,
-                    path_to_expected_png,
-                    diff_dir,
-                )
-
-    if missing_images:
-        print("Missing images:")
-        for i in missing_images:
-            print(i)
-    if mismatched_images:
-        print("Mismatched images:")
-        for i in mismatched_images:
-            print(i)
-
-    # Count summary
-    print(f"Total: {counter}")
-    print(f"Number of missing images: {len(missing_images)}")
-    print(f"Number of mismatched images: {len(mismatched_images)}")
-    print(
-        f"Number of correct images: {counter - len(missing_images) - len(mismatched_images)}"
-    )
-    test_results = Results(diff_dir, task, counter, missing_images, mismatched_images)
-
-    # Make diff_dir readable
-    if os.path.exists(diff_dir):
-        os.system(f"chmod -R 755 {diff_dir}")
-    else:
-        # diff_dir won't exist if all the expected images are missing
-        # That is, if we're in this case, we expect the following:
-        assert len(missing_images) == counter
-
-    return test_results
-
-
-# Multi-machine testing ##########################################################
-
+# Multi-machine testing #########################################################
 # Inspired by https://github.com/E3SM-Project/e3sm_diags/blob/master/docs/source/quickguides/generate_quick_guides.py
